removed statsHelpers into statsController
